Added Characters Editor
